+++ docs/release-notes.md
@@ -1,2 +1,2 @@
-- Added support for hotfix/* branches (configuration will be taken from production).
-- Fixed dependency resolution order to ensure proper recursive download of project dependencies.
+- Added On-Premise cluster provider support.
+- Added On-Premise cluster provider documentation.